requirements.txt
aiogram==3.7.0
httpx==0.27.2
openai==1.54.3
python-dotenv==1.0.1
